+++ .idea/modules.xml
@@ -2,8 +2,8 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/Turtler.iml" filepath="$PROJECT_DIR$/Turtler.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/turtler.iml" filepath="$PROJECT_DIR$/turtler.iml" />
     </modules>
   </component>
 </project>
