+++ app/src/androidTest/java/voyageur/ApplicationTest.java
@@ -1,4 +1,4 @@
-package com.example.skulkarni.turtler;
+package voyageur;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
