+++ app/src/main/java/voyageur/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.skulkarni.turtler;
+package voyageur;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
@@ -8,6 +8,6 @@ public class MainActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(voyageur.R.layout.activity_main);
     }
 }
