+++ app/src/test/java/voyageur/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.skulkarni.turtler;
+package voyageur;
 
 import org.junit.Test;
